set up migrations, hour 1 complete

+++ db/migrate/20170512153131_create_brand.rb
@@ -2,7 +2,7 @@ class CreateBrand < ActiveRecord::Migration[5.1]
   def change
     create_table(:brands) do |t|
        t.column(:name, :string)
-       t.column(:price, :integer)
+       t.column(:price, :decimal)
 
        t.timestamps()
      end
